Broke out Vundle plugin settings into separate file
Vundle plugin settings have been moved to separate file and sourcing has
been moved to the top of the .vimrc file rather than the bottom.

NERDCommenter has been configured to put a space after the comment
character and align the comment character to the left-most of the
block.

Jellybeans theme has been configured to allow the terminal background
colour to be used.